+++ sonavia/UserControls/ActionButton.cs
@@ -1,0 +1,22 @@
+﻿using Guna.UI2.WinForms;
+using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+
+namespace sonavia.UserControls
+{
+    public partial class ActionButton : Guna2Button
+    {
+        public ActionButton()
+        {
+            InitializeComponent();
+        }
+
+    }
+}
