+++ sonavia/Utils/Color.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace sonavia.Utils
+{
+    internal static class Color 
+    {
+        public static System.Drawing.Color orange = ColorTranslator.FromHtml("E78A20");
+        public static System.Drawing.Color lightGray = ColorTranslator.FromHtml("706767");
+        public static System.Drawing.Color darkGray = ColorTranslator.FromHtml("272424");
+        public static System.Drawing.Color fullBlack = ColorTranslator.FromHtml("161616");
+    }
+}
